+++ papermerge/core/apps.py
@@ -1,7 +1,10 @@
 from django.apps import AppConfig
 
 
-class SearchConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'papermerge.search'
-    label = 'search'
+class CoreConfig(AppConfig):
+    name = 'papermerge.core'
+    label = 'core'
+
+    def ready(self):
+        from papermerge.core import checks  # noqa
+        from papermerge.core import signals  # noqa
